requirements.txt
python-dotenv
pika
paho-mqtt
minio
UUID
jsonschema